Misc reorg + skeleton for slot editor
SlotChangeManager is pretty glue-y, might eventually shift to this
implementation

+++ PKHeX.Core/Saves/SAV7.cs
@@ -1230,7 +1230,7 @@ namespace PKHeX.Core
 
         public override bool IsSlotLocked(int box, int slot)
         {
-            if (slot >= 30 || box >= BoxCount)
+            if ((uint)slot >= 30 || (uint)box >= BoxCount)
                 return false;
 
             int slotIndex = slot + (BoxSlotCount * box);
@@ -1239,7 +1239,7 @@ namespace PKHeX.Core
 
         public override bool IsSlotInBattleTeam(int box, int slot)
         {
-            if (slot >= 30 || box >= BoxCount)
+            if ((uint)slot >= 30 || (uint)box >= BoxCount)
                 return false;
 
             int slotIndex = slot + (BoxSlotCount * box);
